+++ include/navkit/api/config/ConfigApi.hpp
@@ -4,7 +4,6 @@
 #pragma once
 
 #include "navkit/api/config/NavKitProductConfigPolicy.hpp"
-#include "navkit/api/config/SensorGraphConfigPolicy.hpp"
 #include "navkit/core/config/ConfigPolicy.hpp"
 #include "navkit/core/estimation/filter/FilterConfigPolicy.hpp"
 #include "navkit/core/estimation/sensor/SensorConfigPolicy.hpp"